adds handleAxiosErrorMessages utility function
