+++ frontend/src/components/constants/errorMessages.tsx
@@ -1,0 +1,2 @@
+export const SIGNUP_FAILED_MESSAGE = "Sign-up failed: "
+export const SIGNIN_FAILED_MESSAGE = "Sign-in failed: "
